CineFilm/requirements.txt
django
djangorestframework
requests
mysqlclient